refactor extractor package
