+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/common/extractor-rule-definition.xml
@@ -2,18 +2,18 @@
 <extractor-rule-definition>
     <extractor-rule id="tableName" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.TableNameExtractor" />
     <extractor-rule id="tableNames" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.TableNamesExtractor" />
-    <extractor-rule id="indexName" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.index.IndexNameExtractor" />
-    <extractor-rule id="indexNames" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.index.IndexNamesExtractor" />
-    <extractor-rule id="columnDefinitions" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.ColumnDefinitionsExtractor" />
-    <extractor-rule id="outlinePrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.constraint.OutlinePrimaryKeyExtractor" />
-    <extractor-rule id="renameTable" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.table.RenameTableExtractor" />
-    <extractor-rule id="renameColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.RenameColumnDefinitionExtractor" />
-    <extractor-rule id="dropColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.DropColumnDefinitionExtractor" />
+    <extractor-rule id="indexName" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.index.IndexNameExtractor" />
+    <extractor-rule id="indexNames" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.index.IndexNamesExtractor" />
+    <extractor-rule id="columnDefinitions" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.ColumnDefinitionsExtractor" />
+    <extractor-rule id="outlinePrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.constraint.OutlinePrimaryKeyExtractor" />
+    <extractor-rule id="renameTable" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.table.RenameTableExtractor" />
+    <extractor-rule id="renameColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.RenameColumnDefinitionExtractor" />
+    <extractor-rule id="dropColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.DropColumnDefinitionExtractor" />
     
-    <extractor-rule id="selectClause" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.SelectClauseExtractor" />
+    <extractor-rule id="selectClause" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.SelectClauseExtractor" />
     <extractor-rule id="fromWhere" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.FromWhereExtractor" />
-    <extractor-rule id="groupBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.GroupByExtractor" />
-    <extractor-rule id="orderBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.OrderByExtractor" />
-    <extractor-rule id="limit" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.LimitExtractor" />
+    <extractor-rule id="groupBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.GroupByExtractor" />
+    <extractor-rule id="orderBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.OrderByExtractor" />
+    <extractor-rule id="limit" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.LimitExtractor" />
     <extractor-rule id="subQueryCondition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.SubQueryConditionExtractor" />
 </extractor-rule-definition>
