+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/oracle/extractor-rule-definition.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <extractor-rule-definition>
-    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.AddColumnDefinitionExtractor"/>
-    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dialect.oracle.definition.column.OracleModifyColumnDefinitionExtractor" />
-    <extractor-rule id="dropPrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dialect.oracle.definition.constraint.OracleDropPrimaryKeyExtractor" />
+    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.AddColumnDefinitionExtractor"/>
+    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.dialect.oracle.OracleModifyColumnDefinitionExtractor" />
+    <extractor-rule id="dropPrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.constraint.dialect.oracle.OracleDropPrimaryKeyExtractor" />
 </extractor-rule-definition>
