+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/postgresql/extractor-rule-definition.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <extractor-rule-definition>
-    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.AddColumnDefinitionExtractor" />
-    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.ModifyColumnDefinitionExtractor" />
-    <extractor-rule id="showParam" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dialect.postgresql.ShowParamExtractor" />
+    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.AddColumnDefinitionExtractor" />
+    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.ModifyColumnDefinitionExtractor" />
+    <extractor-rule id="showParam" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dal.dialect.postgresql.ShowParamExtractor" />
 </extractor-rule-definition>
