+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/sqlserver/extractor-rule-definition.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <extractor-rule-definition>
-    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.AddColumnDefinitionExtractor" />
-    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.definition.column.ModifyColumnDefinitionExtractor" />
-    <extractor-rule id="addPrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dialect.sqlserver.definition.constraint.SQLServerAddPrimaryKeyExtractor" />
+    <extractor-rule id="addColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.AddColumnDefinitionExtractor" />
+    <extractor-rule id="modifyColumnDefinition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.column.ModifyColumnDefinitionExtractor" />
+    <extractor-rule id="addPrimaryKey" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.ddl.constraint.dialect.sqlserver.SQLServerAddPrimaryKeyExtractor" />
     <extractor-rule id="setAutoCommit" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.SetAutoCommitExtractor" />
 </extractor-rule-definition>
